ft_printf 100 %

+++ ft_putptr.c
@@ -6,19 +6,19 @@
 /*   By: izail <marvin@42.fr>                       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/14 18:23:23 by izail             #+#    #+#             */
-/*   Updated: 2021/12/14 18:23:24 by izail            ###   ########.fr       */
+/*   Updated: 2021/12/15 22:14:38 by izail            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-static int ft_len(unsigned int num)
+static int	ft_len(unsigned long long num)
 {
 	int	len;
 
 	len = 0;
-    if (num < 0)
-        len++;
+	if (num < 0)
+		len++;
 	if (num == 0)
 		return (1);
 	while (num != 0)
@@ -36,8 +36,8 @@ static void	ft_put_ptr(unsigned long long num)
 		ft_put_ptr(num / 16);
 		ft_put_ptr(num % 16);
 	}
-    else
-        ft_putchar(LOW_HEX[num]);
+	else
+		ft_putchar(LOW_HEX[num]);
 }
 
 int	ft_putptr(unsigned long long ptr)
